requirements.txt
pytest
boto3
azure-identity
azure-mgmt-authorization
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

